Inline spread syntax
This is especially useful for conditional compilation, where you can now
do things like this:

    {
	a: 1,
	...DEBUG && {
	    b: 2,
	},
	c: 3,
    }

And if DEBUG is true, it’ll compress to this:

    { a: 1, b: 2, c: 3 }

While if DEBUG is false, it’ll compress to this:

    { a: 1, c: 3 }

Finally it is possible to put conditional code mid-object-literal
without any minified size cost!

It works in arrays and function calls too, so that you can do things
like this:

    element.append(
	createElement('h1', null, 'Welcome to X!'),
	...ENVIRONMENT === 'beta' ? [
	    createElement('p', null, 'You’re on beta.'),
	] : [],
	createElement('hr', null),
    );

And the spread will be inlined, so that the function has two or three
simple arguments, depending on the environment.

Fixes #221.