Fix printing of unicode identifiers
This fixes the printing of unicode identifiers.

Important cases here are valid valid identifiers in the Basic Multilingual Plane which are valid to print in ES5+, and Surrogate Pairs which are only valid to print in ES2015+.

This also introduces a new `OutputError` to signal when it's impossible to output an identifier. Mainly, this hits when using a Surrogate Pair variable identifier in ES5.