Ensure default case folding only happens when default is last (#1084)
* Ensure default case folding only happens when default is last

I failed to consider if multiple branches could evaluate to the test condition. Eg,

```js
switch (true) {
    case definitely_true:
    default:
      return 'definitely';
    case maybe_true:
      return 'maybe';
}
```

Here, both `definitely_true` and `maybe_true` can be `=== true`. And we can't eliminate the `case definitely_true:`, because it prevents us from falling into the `maybe_true` branch (when it is true).

* Add test cases for branch reordering