Run Eval removing codeshift
Remove all the places where code is built as a string and then evaled to
make a function. This improves compatability with runtimes where eval is
banned.
The codeshift does the minial set of changes to make this work, and I
have manually fixed a couple of things after running the codeshift.
* Fix the DEFNODE function for the new param order.
* Remove the map/mkshallow functions.

Fixes #925