Switch to GenMapping for sourcemap generation (#1190)
* Switch to GenMapping for sourcemap generation

[`gen-mapping`](https://github.com/jridgewell/gen-mapping) is faster, smaller, and lighter  sourcemap generation library.

This also exposes a new `decodedMap` property on the result object. Decoded maps are free to create (it's a shallow clone of the `GenMapping` instance), and passing them to `@jridgewell/trace-mapping` is copy-free. With Babel  [recently](https://github.com/babel/babel/pull/14497) adding a `decodedMap` field, a dev could pass from the Babel transpilation to Terser without any added memory use for sourcemaps.

And if there's a multi stage build process, a dev could use `@ampproject/remapping` to remap Babel, Terser, and (eg) a bundler's outputs without having to feed input maps into each stage.

* Apply suggestions from code review

Co-authored-by: Fábio Santos <fabiosantosart@gmail.com>

* Fix dependency

* Switch to source-map compatible @jridgewell/source-map

This preserves API compatibility, so that loading source-map in the browser is still supported.

* Update TS types

* Support source-map's async SourceMapConsumer

* Fix sourcesContent check with source-map's SourceMapGenerator

Co-authored-by: Fábio Santos <fabiosantosart@gmail.com>